+++ R/render.R
@@ -345,7 +345,7 @@ render <- function(input,
   }
 
   # check whether this document requires a knit
-  requires_knit <- tolower(xfun::file_ext(input)) %in% c("r", "rmd", "rmarkdown")
+  requires_knit <- tolower(xfun::file_ext(input)) %in% c("r", "rmd", "rmarkdown", "qmd")
 
   # remember the name of the original input document (we overwrite 'input' once
   # we've knitted)
@@ -359,10 +359,10 @@ render <- function(input,
         file_name_without_shell_chars(basename(input)))
 
     if (file.exists(input_no_shell_chars)) {
-      stop("The name of the input file cannot contain the special shell ",
+      stop2("The name of the input file cannot contain the special shell ",
            "characters: ", .shell_chars_regex, " (attempted to copy to a ",
            "version without those characters '", input_no_shell_chars, "' ",
-           "however that file already exists)", call. = FALSE)
+           "however that file already exists)")
     }
     file.copy(input, input_no_shell_chars, overwrite = TRUE)
     intermediates <- c(intermediates, input_no_shell_chars)
@@ -442,7 +442,8 @@ render <- function(input,
   # if this is shiny_prerendered then modify the output format to
   # be single-page and to output dependencies to the shiny.dep file
   shiny_prerendered_dependencies <- list()
-  if (requires_knit && is_shiny_prerendered(front_matter$runtime)) {
+  if (requires_knit && is_shiny_prerendered(front_matter$runtime,
+                                            front_matter$server)) {
 
     # require shiny for the knit
     if (requireNamespace("shiny")) {
@@ -450,7 +451,7 @@ render <- function(input,
         attachNamespace("shiny")
     }
     else
-      stop("The shiny package is required for shinyrmd documents")
+      stop("The shiny package is required for shiny documents")
 
     # source global.R if it exists
     global_r <- file.path.ci(".", "global.R")
@@ -497,8 +498,7 @@ render <- function(input,
 
   # Stop the render process early if the output directory does not exist
   if (!dir_exists(output_dir)) {
-    stop("The directory '", output_dir, "' does not not exist.",
-         call. = FALSE)
+    stop2("The directory '", output_dir, "' does not not exist.")
   }
 
   # use output filename based files dir
@@ -529,7 +529,14 @@ render <- function(input,
   # presume that we're rendering as a static document unless specified
   # otherwise in the parameters
   runtime <- match.arg(runtime)
-  if (identical(runtime, "auto")) runtime <- front_matter$runtime %||% "static"
+  if (identical(runtime, "auto")) {
+    if (is_shiny_prerendered(front_matter$runtime, front_matter$server)) {
+      runtime <- "shiny_prerendered"
+    } else {
+      runtime <- front_matter$runtime %||% "static"
+    }
+  }
+
 
   # set df_print
   context <- render_context()
@@ -569,7 +576,7 @@ render <- function(input,
 
   # determine our id-prefix (add one if necessary for runtime: shiny)
   id_prefix <- id_prefix_from_args(output_format$pandoc$args)
-  if (!nzchar(id_prefix) && is_shiny(runtime)) {
+  if (!nzchar(id_prefix) && is_shiny(runtime, front_matter[["server"]])) {
     id_prefix <- "section-"
     output_format$pandoc$args <- c(output_format$pandoc$args, rbind("--id-prefix", id_prefix))
   }
@@ -723,8 +730,8 @@ render <- function(input,
           inherits(envirParams, "knit_param")
 
         if (!isKnownParamsObject) {
-          stop("params object already exists in knit environment ",
-               "so can't be overwritten by render params", call. = FALSE)
+          stop2("params object already exists in knit environment ",
+               "so can't be overwritten by render params")
         }
       }
 
@@ -788,20 +795,20 @@ render <- function(input,
         identical(tolower(xfun::file_ext(output_file)), "html")))  {
     if (has_html_dependencies(knit_meta)) {
       if (!isTRUE(front_matter$always_allow_html)) {
-        stop("Functions that produce HTML output found in document targeting ",
+        stop2("Functions that produce HTML output found in document targeting ",
              pandoc_to, " output.\nPlease change the output type ",
              "of this document to HTML. Alternatively, you can allow\n",
              "HTML output in non-HTML formats by adding this option to the YAML front",
              "-matter of\nyour rmarkdown file:\n\n",
              "  always_allow_html: true\n\n",
-             "Note however that the HTML output will not be visible in non-HTML formats.\n\n",
-             call. = FALSE)
+             "Note however that the HTML output will not be visible in non-HTML formats.\n\n"
+        )
       }
     }
     if (!identical(runtime, "static")) {
-      stop("Runtime '", runtime, "' is not supported for ",
+      stop2("Runtime '", runtime, "' is not supported for ",
            pandoc_to, " output.\nPlease change the output type ",
-           "of this document to HTML.", call. = FALSE)
+           "of this document to HTML.")
     }
   }
 
@@ -893,22 +900,15 @@ render <- function(input,
       # in case the output format turns on the --file-scope flag, run its
       # file_scope function to split the input into multiple files
       input_files <- input
-      if (!is.null(output_format$file_scope) &&
-          length(inputs <- output_format$file_scope(input)) > 1) {
-
-        # add the --file-scope option
-        pandoc_args <- c(pandoc_args, "--file-scope")
-
-        # write the split content into *.split.md files
-        input_files <- unlist(lapply(inputs, function(input) {
-          file <- file_with_meta_ext(input$name, "split", "md")
-          file <- file.path(dirname(input), file)
-          write_utf8(input$content, file)
-          file
-        }))
-
-        # cleanup the split files after render
-        on.exit(unlink(input_files), add = TRUE)
+      if (is.function(output_format$file_scope)) {
+        input_files <- file_scope_split(input, output_format$file_scope)
+        # ignore if input_files has not really been splitted
+        if (length(input_files) > 1) {
+          # add the --file-scope option
+          pandoc_args <- c(pandoc_args, "--file-scope")
+          # cleanup the split files after render
+          on.exit(unlink(input_files), add = TRUE)
+        }
       }
 
       # if we don't detect any invalid shell characters in the
@@ -966,7 +966,7 @@ render <- function(input,
     # if the output format is LaTeX, first convert .md to .tex, and then convert
     # .tex to .pdf via latexmk() if PDF output is requested (in rmarkdown <=
     # v1.8, we used to call Pandoc to convert .md to .tex and .pdf separately)
-    if (output_format$pandoc$keep_tex || knitr::is_latex_output()) {
+    if (output_format$pandoc$keep_tex || pandoc_to %in% c("latex", "beamer")) {
       # do not use pandoc-citeproc if needs to build bibliography
       convert(texfile, run_citeproc && !need_bibtex)
       # patch the .tex output generated from the default Pandoc LaTeX template
@@ -1166,8 +1166,8 @@ resolve_df_print <- function(df_print) {
     else if (df_print == "default")
       df_print <- print
     else
-      stop('Invalid value for df_print (valid values are ',
-           paste(valid_methods, collapse = ", "), call. = FALSE)
+      stop2('Invalid value for df_print (valid values are ',
+           paste(valid_methods, collapse = ", "))
   }
 
   df_print
@@ -1194,3 +1194,21 @@ resolve_df_print <- function(df_print) {
 #' @keywords NULL
 #' @export
 output_metadata = knitr:::new_defaults()
+
+file_scope_split <- function(input, fun) {
+  inputs <- fun(input)
+
+  # file_scope_fun should split the input file in several
+  # do nothing if not and return input file unsplited
+  if (length(inputs) <= 1) return(input)
+
+  # write the split content into *.split.md files
+  input_files <- lapply(inputs, function(f) {
+    file <- file_with_meta_ext(f$name, "split", "md")
+    file <- file.path(dirname(input), file)
+    write_utf8(f$content, file)
+    file
+  })
+
+  unlist(input_files)
+}